# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AqarGo API",
        "description": "AqarGo real estate platform API documentation",
        "contact": {
            "email": "support@aqargo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmad"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmad@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "fcm_token": {
                                        "type": "string",
                                        "example": "fcm_token_here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and get JWT token",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmad@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123",
                                        "minLength": 8
                                    },
                                    "fcm_token": {
                                        "type": "string",
                                        "example": "fcm_token_here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "401": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the authenticated user",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "401": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh the JWT token",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "401": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user's info",
                "operationId": "f0b15943f55143d36f606a2dbd1d2366",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "401": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/forgetPassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send password reset code to email",
                "operationId": "a596bf0086d25bea6b3721256d06add1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmad@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            }
        },
        "/api/auth/checkCode": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify the password reset code",
                "operationId": "715451e030d64688fd5ab7c760063d85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmad@example.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            }
        },
        "/api/auth/resetPassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset the user's password",
                "operationId": "ed3326bc53cbcc940f34d8b4048a745d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmad@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation Error."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT token"
            }
        }
    },
...
}